Move Room and walls to gameobject package
